% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{center}
\alias{center}
\title{Mean-center a variable}
\usage{
center(x)
}
\arguments{
\item{x}{Numeric vector}
}
\description{
This is a convenience function equal to scale(x, scale=FALSE)
}
\examples{
data(anchoring)
center(anchoring$everest_feet)
}
