% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cheerleader}
\alias{cheerleader}
\title{Data from Experiment 1 of Carragher, D.J., Thomas, N.A., Gwinn, O.S. et al. (2019) Limited evidence of hierarchical encoding in the cheerleader effect. Scientific Reports, 9, 9329. https://doi.org/10.1038/s41598-019-45789-6}
\format{
An object of class \code{data.frame} with 192 rows and 9 columns.
}
\source{
\url{https://osf.io/je5u7/}. Carragher, D.J., Thomas, N.A., Gwinn, O.S. et al. (2019) Limited evidence of hierarchical encoding in the cheerleader effect. Scientific Reports, 9, 9329 \doi{10.1038/s41598-019-45789-6}.
}
\usage{
cheerleader
}
\description{
\@format A data frame with 320 observations of 16 variables:
\describe{
\item{Participant}{(factor) Participant ID}
\item{Age}{(numeric) Participant age in years}
\item{Sex}{(factor) Participant sex (Male or Female)}
\item{Task}{(factor) Identical-Distractors, or Self-Distractors.}
\item{LineClickAccuracy}{Measure of average response deviation from the visual analogue scale; scores > +/- 2.00 constitute exclusion.}
\item{Excluded}{(numeric) Indicator whether participant was excluded from main analysis (0 = no, 1 = yes)}
\item{WhyExcluded}{(character) explanation for exclusion}
\item{Item}{(factor) Item description}
\item{Response}{Attractiveness rating for the target face on a visual analogue scale ranging from “Very Unattractive” (0) to “Very Attractive” (100)}
}
}
\keyword{datasets}
