% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_Anova.R
\name{expand_Anova}
\alias{expand_Anova}
\title{Expand all contrast terms in car::Anova}
\usage{
expand_Anova(mod, type = c("III", "II", 3, 2), ...)
}
\arguments{
\item{mod}{A model of class lm}

\item{type}{SS Type (see ?car::Anova)}

\item{...}{Further arguments passed to Anova}
}
\value{
Object of class anova returned by car::Anova
}
\description{
This is an experimental function
}
\examples{
data("tetris2015")
mod <- lm(Days_One_to_Seven_Number_of_Intrusions ~ Condition, data=tetris2015)
car::Anova(mod,type=3) # default type III Anova table
expand_Anova(mod,type=3)
}
