% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fifa2010}
\alias{fifa2010}
\title{Predictions by Paul the Octopus in the 2010 Fifa World Cup.}
\format{
A data frame with 8 rows and 4 variables:
\describe{
\item{Match}{countries playing}
\item{Prediction}{country predicted to win}
\item{Result}{score at the end of the match}
\item{Outcome}{whether Paul was correct or incorrect}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Paul_the_Octopus}
}
\usage{
fifa2010
}
\description{
A dataset containing the predictions and outcomes of matches
in the 2010 FIFA European Cup.
}
\keyword{datasets}
