% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gestures}
\alias{gestures}
\title{Data from Winter, B., & Burkner, P. (2021) Poisson regression for linguists: A tutorial introduction to modelling count data with brms. Language and Linguistics Compass, 15, e12439 \doi{10.1111/lnc3.12439}}
\format{
An object of class \code{data.frame} with 54 rows and 6 columns.
}
\source{
\url{https://osf.io/6j8kc}.
}
\usage{
gestures
}
\description{
\@format A data frame with 54 observations of 6 variables:
\describe{
\item{ID}{(factor) Participant ID}
\item{context}{(factor) Whether talking to a friend or professor}
\item{duration}{(numeric) Duration of the interaction}
\item{language}{(factor) Language spoken: Catalan or Korean}
\item{gender}{(factor) Participant gender (M = male, F = female)}
\item{gestures}{(numeric) number of gestures in the interaction.}
}
}
\keyword{datasets}
