% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{legacy2015}
\alias{legacy2015}
\title{Legacy motives and pro-environmental behaviour}
\format{
A data frame with 245 rows and 9 variables:
\describe{
\item{id}{(numeric) ID variable relating to the original dataset}
\item{sex}{(character) biological sex of participant (male or female)}
\item{age}{(numeric) age in years}
\item{legacy}{(numeric) Sverage of 8 items reflecting legacy motivation, on a scale from 1 (Not at all) to 6 (A great amount)}
\item{belief}{(numeric) average of 5 items reflecting belief in climate change, on a scale from 1 (Strongly Disagree) to 7 (Strongly Agree)}
\item{intention}{(numeric) average of 8 items reflecting intention to act in a pro-environmental way, on a scale from 1 (Never) to 6 (All the time)}
\item{education}{(numeric) Level of education, 1 = 8th grade or less, 2 = Some high school, 3 = Graduated high school, 4 = Some college or technical school, 5 = Graduated college or technical school, 6 = Post-graduate}
\item{income}{(numeric) Approximate household income, 1 = less than $20K, 2 = $20K-$35K, 3 = $35K-$50K 4 = $50K-$75K, 5 = $75K-100K, 6 = more than 100K}
\item{donation}{(numeric) Donation of possible bonus payment, between $0 and $10}
}
}
\source{
Harvard DataVerse \url{https://dataverse.harvard.edu/dataset.xhtml?persistentId=doi:10.7910/DVN/27740&version=1.0}
}
\usage{
legacy2015
}
\description{
Legacy motives and pro-environmental behaviour
}
\examples{
\dontrun{
 # this dataset was processed from the raw data as follows:
 tdat <- read.csv("legacy study - pilot_data.csv")
}
}
\keyword{datasets}
