% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sd.R
\name{sample_sd}
\alias{sample_sd}
\title{Compute the sample standard deviation}
\usage{
sample_sd(x, na.rm = FALSE)
}
\arguments{
\item{x}{Numeric vector}

\item{na.rm}{(logical) Should missing values be removed?}
}
\description{
Compute the sample standard deviation
}
\examples{
data(anchoring)
sample_sd(anchoring$everest_feet)
}
