% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_var.R
\name{sample_var}
\alias{sample_var}
\title{Compute the sample variance}
\usage{
sample_var(x, na.rm = FALSE)
}
\arguments{
\item{x}{Numeric vector}

\item{na.rm}{(logical) Should missing values be removed?}
}
\description{
Compute the sample variance
}
\examples{
data(anchoring)
sample_var(anchoring$everest_feet)
}
