% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{speeddate}
\alias{speeddate}
\title{Speed dating}
\format{
A data frame with 1562 rows and 32 variables:
\describe{
\item{iid}{(numeric) unique ID variable of participant}
\item{pid}{(numeric) unique ID variable of date partner}
\item{gender}{(character) gender of participant (male or female)}
\item{age}{(numeric) age in years}
\item{date_like}{(numeric) how much they like their date partner in general (between  1 and 10)}
\item{other_like}{(numeric) how much their date partner likes them (between  1 and 10)}
\item{date_want}{do they want to go on another date with their date partner? (1 = yes, 0 = no)}
\item{other_want}{does their date partner want to go on another date with them? (1 = yes, 0 = no)}
\item{match}{do they both want to go on another date with each other? (1 = yes, 0 = no)}
\item{self_attr}{how attractive do they think they are? (between 1 and 10)}
\item{self_sinc}{how sincere do they think they are? (between 1 and 10)}
\item{self_intel}{how intelligent do they think they are? (between 1 and 10)}
\item{self_fun}{how much fun do they think they are? (between  1 and 10)}
\item{self_amb}{how ambitious do they think they are? (between 1 and 10)}
\item{other_attr,other_sinc,other_intel,other_fun,other_amb}{how attractive etc does their date partner think they are? (between  1 and 10)}
\item{other_shar}{how much does their date partner think they share hobbies and interests? (between 1 and 10)}
\item{date_attr,date_sinc,date_intel,date_fun,date_amb,date_shar}{how do they rate their date partner's attractiveness etc? (between 1 and 10)}
\item{self_imp_attr,self_imp_sinc,self_imp_intel,self_imp_fun,self_imp_amb,self_imp_shar}{how important do they find attractiveness etc in a partner? (between 1 and 10)}
\item{other_imp_attr,other_imp_sinc,other_imp_intel,other_imp_fun,other_imp_amb,other_imp_shar}{how important does their date partner find attractiveness etc? (between 1 and 10)}
}
}
\source{
Kaggle \url{https://www.kaggle.com/annavictoria/speed-dating-experiment}
}
\usage{
speeddate
}
\description{
A subset of cases (wave 6-9) and variables (see below) from an experiment on speed dating.
by Columbia Business School professors Ray Fisman and Sheena Iyengar for their paper
Gender Differences in Mate Selection: Evidence From a Speed Dating Experiment.
}
\keyword{datasets}
