% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{trump2016}
\alias{trump2016}
\title{Trump votes in 2016 for 51 US states}
\format{
A data frame with 4632 rows and 5 variables:
\describe{
\item{state}{Name of the state}
\item{hate_groups}{Number of hate groups in the state in 2016 as reported by the Southern Poverty Law Center (\url{https://www.splcenter.org/hate-map})}
\item{population}{Number of citizens in the state in 2016}
\item{hate_groups_per_million}{Number of hate groups per million citizens}
\item{percent_bachelors_degree_or_higher}{Percentage of citizens with a bachelor's degree of higher}
\item{percent_in_poverty}{Percentage of citizens below the poverty threshold}
\item{percent_Trump_votes}{Percentage of votes for Trump in the 2016 elections}
}
}
\source{
CSI Without Dead Bodies "Hate Groups and Trump's Vote\%: Predictive effect present when education and poverty are considered" \url{https://www.csiwithoutdeadbodies.com/2017/02/hate-groups-and-trumps-vote-predictive.html}
}
\usage{
trump2016
}
\description{
Trump votes in 2016 for 51 US states
}
\keyword{datasets}
