% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.R
\name{center}
\alias{center}
\title{Mean-centered values}
\usage{
center(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
A numeric vector with mean-centered values
}
\description{
\code{center} computes mean-centered values. It is a convenience wrapper to \code{scale}, equal to scale(x, scale=FALSE)
}
\examples{
data(anchoring)
center(anchoring$everest_feet)
}
