% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_flat_violin.R
\name{geom_flat_violin}
\alias{geom_flat_violin}
\title{Half violin plot}
\source{
url{https://gist.github.com/dgrtwo/eb7750e74997891d7c20}
}
\usage{
geom_flat_violin(
  mapping = NULL,
  data = NULL,
  stat = "ydensity",
  position = "dodge",
  trim = TRUE,
  scale = "area",
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{The mapping}

\item{data}{data.frame}

\item{stat}{statistic (don't change)}

\item{position}{position dodge}

\item{trim}{Logical}

\item{scale}{Scale (don't change)}

\item{show.legend}{Logical}

\item{inherit.aes}{Logical}

\item{...}{other arguments}
}
\description{
Half violin plot
}
\examples{
library(ggplot2)
data(diamonds)
ggplot(diamonds, aes(cut, carat)) + geom_flat_violin() + coord_flip()
}
