% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_qq_marginals.R
\name{plot_qq_marginals}
\alias{plot_qq_marginals}
\title{A Q-Q plot with histograms in the marginals}
\usage{
plot_qq_marginals(
  x,
  newpage = TRUE,
  xlab = "Observed Quantiles",
  ylab = "Theoretical quantiles",
  xlim = grDevices::extendrange(c(min(x), max(x))),
  ylim = NULL,
  main = NULL,
  sub = NULL,
  axes = TRUE,
  border = TRUE,
  ...
)
}
\arguments{
\item{x}{A numeric vector}

\item{xlab}{Label for x-axis}

\item{ylab}{Label for y-axis}

\item{xlim}{range of x values shown}

\item{ylim}{range of y values shown}

\item{main}{Main title}

\item{sub}{Subtitle}

\item{axes}{(logical) Draw axes?}

\item{border}{(logical) Draw a border?}

\item{...}{Further arguments}
}
\description{
A Q-Q plot with histograms in the marginals
}
\examples{
data(anchoring)
plot_qq_marginals(anchoring$everest_feet)
}
