% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raincloud.R
\name{plot_raincloud}
\alias{plot_raincloud}
\title{Create a raincloud plot}
\source{
Allen M, Poggiali D, Whitaker K et al. Raincloud plots: a multi-platform tool for robust data visualization. Wellcome Open Res 2019, 4:63 (\url{https://doi.org/10.12688/wellcomeopenres.15191.1})
}
\usage{
plot_raincloud(data, y, horizontal = FALSE, groups, point_size = 0.5, ...)
}
\arguments{
\item{data}{Data.frame (or tibble)}

\item{y}{The unquoted name of the variable in data for which to create the raincloud plot}

\item{horizontal}{(logical) change the orientation of the plot}

\item{groups}{An unquoted name of grouping variable in data (ideally a factor)}

\item{point_size}{Size of the jittered points}

\item{...}{Other arguments, passed to ggplot(aes(...))}
}
\description{
Create a raincloud plot
}
\examples{
data(anchoring)
plot_raincloud(anchoring,y=everest_feet)
plot_raincloud(anchoring,y=everest_feet,groups=anchor)
plot_raincloud(anchoring,y=everest_feet,groups=anchor) + 
    ggplot2::facet_wrap(~us_or_international) + 
    ggplot2::ylab("How high is Mount Everest (in feet)?")
}
