% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{sample_mode}
\alias{sample_mode}
\title{Compute a sample mode}
\usage{
sample_mode(x)
}
\arguments{
\item{x}{Numeric vector}
}
\description{
Compute a sample mode
}
\examples{
data(anchoring)
sample_mode(anchoring$everest_feet)
# Multiple modes give a warning:
sample_mode(c(1,1,1,2,2,2))
}
